backend/requirements.txt
Flask==2.1.0
Flask-SQLAlchemy>=2.5.1
SQLAlchemy<2.0
Flask-Migrate==3.1.0
Flask-JWT-Extended==4.4.0
PyMySQL==1.0.2
Werkzeug==2.1.0
python-dotenv
flask-cors
